+++ trunk/xrGame/holder_custom.h
@@ -21,7 +21,7 @@ public:
 	virtual	void			UpdateEx			(float fov){}; //called by owner
 	virtual CHolderCustom	*cast_holder_custom	()				{return this;}
 			bool			Engaged				()				{return m_owner!=NULL;}
-	virtual void			OnMouseMove			(int x, int y)	= 0;
+	virtual void OnMouseMove(float x, float y) = 0;
 	virtual void			OnKeyboardPress		(int dik)		= 0;
 	virtual void			OnKeyboardRelease	(int dik)		= 0;
 	virtual void			OnKeyboardHold		(int dik)		= 0;
